Add a `wait` state to the particle state machine
Wait allows for external control of the particle's execution which
will keep the counting of iterations and the other termination logic
separated from the logic of moving, accelerating, and evaluating a
particle. Does create a bit of a concern as to whether we are
maximizing the available parallelism by having all particles update
concurrently, but I don't believe this is much of a concern since the
computationally expensive step (evaluating the objective function) is
executed concurrently by all particles.

Comments and documentation is also substantially improved, and some
convenient type aliases are added.