Refactor position reporting and accelerate state
Rather than reporting position and value to neighbors on state entry
we report it when we leave the eval state. This is a much better fit
with the event-driven state machine model and results in much cleaner
code.

Additionally, the accelerate state has been refactored to use a helper
function that adds support for particles with no neighbors. As I read
the diff, however, the refactoring won't work since the helper
function will attempt to trigger a state transition out of the enter
event which is not allowed (will cause the statem process to
crash). I'll need to put a little more thought in to how to fix this.