Rebuilt particle module as a datastructure only
Remove the gen_statem implementation and build a simple stand-alone
version of the particle code. This sets up a much better separation of
concerns between the PSO algorithm and the process structure that will
be built to run it. Here we have only the PSO algorithm. Next we can
add:

- tests

- code to create/run a sequential PSO instance (one process) as a step
  towards the parallel version. I think for now writing a sequential
  version will let me see what kind of API the parallel version will
  need to support.