Refactor the particle module
The statem approach for particles was getting ahead of itself. This
refactor removes all the OTP parts and leaves only the core PSO
algorithm. I will be adding a separte 'particle_statem' module (might
actually be a gen_server, not a statem) that uses the particle module.
This is a much better separation of concerns, the core algorithm can
be easily tested & modified since it isn't tied to process and we can
call in to it directly.

A bit of work still remains: I would like to split out the vector math
into its own module to keep things tidier, also need to support
parameters for the PSO algorithm (notably limits on the problem domain
and on the maximum velocity, but extensible to support other params).

This also adds a single API function to the 'pso' module, which can be
use to run a PSO instance sequentially. Will be modified/extended to
run the parallel version when it is implemented.