Make particle_server:eval/2 a gen_server:call
Using a call instead of a cast lets us return an indication of how
many total iterations will have been completed once the iteration is
done (for example, if the server has already completed some
iterations), and it lets us signal to the caller if the server is
already iterating.

Add particle_server:stop/1 to the API.

Add particle_server:state/2 to the API. This lets us block the caller
until some number of iterations have been completed. It requires
adding another field to the server state to store requests that have
not been completed yet. The stored requests are scanned each time an
iteration is completed.

Bug fixed: assign list of neigbors _before_ trying to send the initial
state to all neighbors when handling the {assing_neighbors, _} cast.