Correct calculation of global acceleration in particle:accelerate/4
The global acceleration vector was being computed incorrectly. The
acceleration vector should be aligned with the vector from particle's
current position to the best neighbor's position. In the incorrect
implementation, the global acceleration vector was aligned with the
vector from the best neighbor's position to the particle's "personal
best" position.