Make particle_server tests more robust
Because of the stochastic nature of the PSO algorithm, Sometimes
the final position does not end up where we expect it to with only
two particles. To make random test failures caused by the stochastic
nature of the algorithm less likely I have changed the objective
function from abs/1 to a quadratic, and reduced the maximum
velocity (which should reduce the likelyhood of overshooting the
minimum and prevent large oscilations when overshoot does
happen).

Also doubled the number of iterations used in the test. This costs a
little time (roughly an extra 0.01 seconds), but for only one test I
think it is reasonable.